+++ app/src/main/java/com/example/osu_bathroom_app/map/MapFragment.java
@@ -1,4 +1,4 @@
-package com.example.osu_bathroom_app;
+package com.example.osu_bathroom_app.map;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,6 +7,7 @@ import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
+import com.example.osu_bathroom_app.R;
 import com.mapbox.maps.MapView;
 import com.mapbox.maps.Style;
