+++ app/src/main/java/com/example/osu_bathroom_app/models/BathroomListViewModel.java
@@ -1,4 +1,4 @@
-package com.example.osu_bathroom_app;
+package com.example.osu_bathroom_app.models;
 
 import android.util.Log;
