+++ app/src/main/java/com/example/osu_bathroom_app/models/BathroomRepository.java
@@ -1,4 +1,4 @@
-package com.example.osu_bathroom_app;
+package com.example.osu_bathroom_app.models;
 
 import android.util.Log;
 
@@ -68,8 +68,6 @@ public class BathroomRepository
                         }
                     });
                 }
-
-
             }
