+++ app/src/main/res/layout/activity_main.xml
@@ -6,4 +6,4 @@
     android:name="com.example.HomePageFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity" />
+    tools:context=".main.MainActivity" />
