+++ app/src/main/res/layout/fragment_add_bathroom.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddBathroomFragment">
+    tools:context=".list_page.AddBathroomFragment">
 
     <EditText
         android:id="@+id/editTextName"
