+++ app/src/main/res/layout/fragment_bathroom_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/list_container"
-    tools:context=".BathroomListFragment">
+    tools:context=".list_page.BathroomListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/bathroom_list"
