+++ app/src/main/res/layout/fragment_home_page.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomePageFragment">
+    tools:context=".main.HomePageFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
